requirements.txt
fastapi
uvicorn
azure-storage-blob
azure-identity